+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 ## vNext
 
+**BREAKING CHANGES**
+- Fix `--principle-id` flag to be `principal-id`
 - Remove MasterAccountCreds from dce.yml. Use default AWS CLI creds instead
 - Indent JSON output from CLI commands